requirements.txt
# AI LinkedIn Machine - Dependencies

# RSS Feed Ingestion
feedparser>=6.0.0

# Date parsing
python-dateutil>=2.8.0

# OpenAI API for summarization
openai>=1.0.0

# Image Optimization
Pillow>=10.0.0

# Web App (for mobile access)
streamlit>=1.28.0

# Optional: Enhanced drag-and-drop support for desktop GUI
# tkinterdnd2>=0.3.0  # pip install tkinterdnd2 (optional, for drag-drop)

# Optional: Enhanced image processing
# opencv-python>=4.8.0  # For advanced face detection in smart crop (optional)
